Move library code into dedicated folder
