+++ build.gradle
@@ -1,21 +1,3 @@
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion 28
-    defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 28
-        versionCode 3
-        versionName '0.3'
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
 buildscript {
     repositories {
         jcenter()
@@ -26,14 +8,10 @@ buildscript {
     }
 }
 
-repositories {
-    jcenter()
-    mavenCentral()
-    google()
-}
-
-dependencies {
-    implementation 'com.android.support:support-annotations:28.0.0'
-    implementation 'com.arasthel:asyncjob-library:1.0.3'
-    implementation 'com.bluelinelabs:logansquare:1.0.6'
+allprojects {
+    repositories {
+        jcenter()
+        mavenCentral()
+        google()
+    }
 }
