+++ library/src/main/AndroidManifest.xml
