+++ library/src/main/java/com/peak/salut/Callbacks/SalutDeviceCallback.java
@@ -1,0 +1,7 @@
+package com.peak.salut.Callbacks;
+
+import com.peak.salut.SalutDevice;
+
+public interface SalutDeviceCallback {
+    void call(SalutDevice device);
+}
