+++ library/src/main/java/com/peak/salut/SalutDevice.java
