+++ library/src/main/res/values/strings.xml
