+++ settings.gradle
@@ -1,0 +1,3 @@
+include ':library'
+
+
