--- src/main/java/com/peak/salut/Callbacks/SalutDeviceCallback.java
@@ -1,7 +1,0 @@
-package com.peak.salut.Callbacks;
-
-import com.peak.salut.SalutDevice;
-
-public interface SalutDeviceCallback {
-    void call(SalutDevice device);
-}
